canonical-json: Move redaction functions and types to separate module
